requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
email-validator>=2.2.0,<3.0.0
requests>=2.31.0,<3.0.0
